test: move the framework SDK stubs into conftest
They were a module three test files imported by hand. conftest is where
pytest already looks, so the separate file bought an import line and a file.

This does not change the file count, which I claimed in the first draft of
this message before checking: deleting the module drops one file but conftest
becomes modified, so the diff stays at 300. That is at GitHub's rendering
limit rather than past it. Keeping the change because it is the right home
for the stubs, not because it buys headroom.

agent-os 2923 passed.

Copilot-Session: 384bda36-161f-4689-9b2c-0a6d5c90776f
