+++ agent-governance-python/agent-os/tests/test_adapter_enforcement.py
@@ -20,7 +20,7 @@ from typing import Any
 
 import pytest
 
-from _framework_stubs import install as _install_framework_stubs
+from conftest import install as _install_framework_stubs
 
 _install_framework_stubs()
