+++ agent-governance-python/agent-os/tests/test_adapter_interception.py
@@ -17,8 +17,8 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from _framework_stubs import DropMessage, FunctionCall
-from _framework_stubs import install as _install_framework_stubs
+from conftest import DropMessage, FunctionCall
+from conftest import install as _install_framework_stubs
 
 _install_framework_stubs()
