+++ agent-governance-python/agent-os/tests/test_governed_clients.py
@@ -16,7 +16,7 @@ from unittest.mock import MagicMock
 
 import pytest
 
-from _framework_stubs import install as _install_framework_stubs
+from conftest import install as _install_framework_stubs
 
 _install_framework_stubs()
